% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msy.R
\name{msy}
\alias{msy}
\title{Calculate max sustainable yield}
\usage{
msy(k, r)
}
\arguments{
\item{k}{is the carrying capacity (individuals)}

\item{r}{intrinsic growth rate (1/yr)}
}
\value{
Returns the maximum sustainable harvest for a fischery
}
\description{
Calculate max sustainable yield
}
\examples{
msy(300, 0.06)
}
